Count only committed transactions when calculating the next nonce.
Before this commit we would have a situation when we query the node for
the number of transactions (including pending) and use this number as
a nonce.

Example:

Committed transactions:
0, 1, 2

Pending transactions:
3

Using "pending" defaultBlock we get 4 and so the next nonce will be 4.
Using "latest" defaultBlock we get 3 and so the next nonce will be 3.

In this case using "latest" will be better as this way we can (possibly)
replace the pending transaction with the updated one.